+++ src/element-auto.ts
@@ -1,0 +1,6 @@
+import { defineCodexChatElement } from "./element";
+
+defineCodexChatElement();
+
+export { defineCodexChatElement } from "./element";
+export type { DefineCodexChatElementOptions } from "./element";